+++ NEWS.md
@@ -1,3 +1,9 @@
+# patternogram 0.1.2
+
+* Adds an option of splitting the patternograms by raster layers
+* Adds a new function c.patternogram() to merge many patternogram objects
+* Updates autoplot to handle many patternograms at the same time
+
 # patternogram 0.1.1
 
 * Adds parallalization possibility when calculating uncertainty intervals